# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "uSocial API",
        "description": "API documentation for uSocial",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chat/unlock": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Unlock chat with an influencer",
                "description": "Deduct gems from the user and unlock chat with the specified influencer.",
                "operationId": "4a13d5b6b3f7748a5fa67d460d240612",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "influencer_id"
                                ],
                                "properties": {
                                    "influencer_id": {
                                        "description": "ID of the influencer to unlock chat with",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat unlocked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat unlocked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not enough gems or chat already unlocked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not enough gems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Influencer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Influencer not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}